+++ source/extensions/isaacsim.core.deprecation_manager/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [0.3.3] - 2026-02-07
+### Changed
+- Remove deprecated asset browser settings
+
 ## [0.3.2] - 2025-11-07
 ### Changed
 - Removed code to enable the `omni.isaac.ml_archive` extension when importing PyTorch via `import_module`